Move duplicate check to middleware
