+++ src/controllers/PartyController.js
@@ -4,15 +4,6 @@ import PartyModel from '../models/PartyModel';
 class PartyController {
   static create(req, res) {
     const { name, hqAddress, logoUrl } = req.body;
-    const partyExists = PartyModel.find(party => (
-      party.name.toLowerCase() === name.toLowerCase()
-    ));
-    if (partyExists) {
-      return res.status(400).json({
-        status: 400,
-        error: 'The party already exists',
-      });
-    }
     const newParty = {
       id: PartyModel.length + 1,
       name,
