+++ src/middlewares/PartyValidator.js
@@ -1,3 +1,5 @@
+import PartyModel from '../models/PartyModel';
+
 class PartyValidator {
   static createPartyValidator(req, res, next) {
     req.check('name', 'The party name is required').notEmpty();
@@ -14,11 +16,20 @@ class PartyValidator {
         errors: validationErrors,
       });
     }
+    const partyExists = PartyModel.find(party => (
+      party.name.toLowerCase() === req.body.name.toLowerCase()
+    ));
+    if (partyExists) {
+      return res.status(400).json({
+        status: 400,
+        error: 'The party already exists',
+      });
+    }
     const { name, hqAddress, logoUrl } = req.body;
     req.body.name = name.trim();
     req.body.hqAddress = hqAddress.trim();
     req.logoUrl = logoUrl.trim();
-    next();
+    return next();
   }
 }
